requirements.txt
coverage==4.5.1
docopt==0.6.2
grpcio-tools==1.11.0
grpcio==1.11.0
protobuf==3.5.2.post1
protol==0.0.3
pytest==3.5.0
python-dotenv==0.8.2
scrypt=0.8.6
